pronto pre lint

+++ user-api/src/services/user.js
@@ -2,22 +2,36 @@ const UserRepository = require('../repositories/user')
 const { ErrorHandler } = require('../middlewares/error-handler')
 
 exports.create = async (user) => {
-  console.log('>>>chegou')
   const createdUser = await UserRepository.create(user)
   return createdUser
 }
 
-exports.update = async (id, user) => {
+exports.update = async (userAuthId, id, user) => {
+  if (userAuthId !== id) {
+    throw new ErrorHandler(403, ['Forbidden'])
+  }
+
   const updatedUser = await UserRepository.update(id, user)
+
   return updatedUser
 }
 
-exports.findById = async (companyId) => {
-  const user = await UserRepository.findById(companyId)
-  return user
+exports.findById = async (userAuthId, id) => {
+  if (userAuthId !== id) {
+    throw new ErrorHandler(403, ['Forbidden'])
+  }
+
+  const foundUser = await UserRepository.findById(id)
+
+  return foundUser
 }
 
-exports.delete = async (id) => {
+exports.delete = async (userAuthId, id) => {
+  if (userAuthId !== id) {
+    throw new ErrorHandler(403, ['Forbidden'])
+  }
+
   const deletedUser = await UserRepository.delete(id)
+
   return deletedUser
 }
